No bugs
There are no bugs in the code but still the results of the BP vs exact marginals on single nodes and of Z are not the same. Areasone for this could be because I am using MF approximation over the beliefs which means that the joint of the system is the product of all the single node beliefs.